pyproject.toml
[project]
name = "NiceStreams"
version = "0.1.0"
description = "ai notes"
authors = [
    {name = "ShivaNKA0", email = "shiva.kanneboina@gmail.com"},
]
dependencies = [
    "Flask>=3.0.3",
    "google-api-python-client>=2.149.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "tailwindcss>=0.0.1",
    "openai>=1.51.2",
    "PyPDF2>=3.0.1",
    "env>=0.1.0",
    "python-env>=1.0.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
